Fix 'Auto-scale images to viewport' to exclude hidden items
You can use the advanced options to hide values on an axis, but the code was calculating the total width assuming that all items will be visible.

So if you had 10 elements on an axis and only 5 were to be shown, it would still scale stuff down to fit all 10 items, making the images too small.