Made the navigation header stick to the left
Just chaning top_nav_bar to be sticky won't do anything because its
container (the `<center>` tag) was only as wide as the viewport. So even
if it was sticky-left, its parent center tag would get scrolled right
off the screen anyway. The table is wider than the viewport so it
gets scrolled properly.

The structure of stuff is:

- `<center>` (now `div#content_box`)
   - top_nav_bar
   - `<div>` (is now `#image_table_box`)
     - table

This commit makes `image_table_box` an inline-block to make it take the
width of the table instead of the width of its parent, which was just
the viewport width. It also makes content_box an inline-block so it
takes the width of image_table_box instead of the viewport width.

Now top_nav_bar's container is the width of the table instead of the
viewport, so now making it sticky-left works.

The min-width of image_table_box is 100% (at least viewport width) so
that the table is centered if its content isn't wide enough.

Also made the save_image_helper have the secondary class so it has an
opaque background.